Move to toplevel, split tests
